# backend/src/main/java/com/shipflow/order/api/ShipmentOrderManagementController.java
package com.shipflow.order.api;
import com.shipflow.common.api.ApiResponse; import com.shipflow.order.api.model.*; import com.shipflow.order.application.*; import jakarta.validation.Valid; import org.springframework.beans.factory.annotation.Autowired; import org.springframework.http.*; import org.springframework.security.core.annotation.AuthenticationPrincipal; import org.springframework.security.oauth2.jwt.Jwt; import org.springframework.web.bind.annotation.*; import java.time.LocalDateTime; import java.util.List;
@RestController @RequestMapping("/api/v1/orders") public class ShipmentOrderManagementController {private final ShipmentOrderQueryApplicationService query;private final ShipmentOrderLifecycleApplicationService lifecycle;private final ShipmentOrderExportApplicationService exporter;public ShipmentOrderManagementController(ShipmentOrderQueryApplicationService q,ShipmentOrderLifecycleApplicationService l){this(q,l,null);}@Autowired public ShipmentOrderManagementController(ShipmentOrderQueryApplicationService q,ShipmentOrderLifecycleApplicationService l,ShipmentOrderExportApplicationService exporter){query=q;lifecycle=l;this.exporter=exporter;}
 @GetMapping public ApiResponse<ShipmentOrderPageResponse> list(@RequestParam(required=false)String orderNo,@RequestParam(required=false)String status,@RequestParam(required=false)String workbenchFilter,@RequestParam(required=false)Long storeId,@RequestParam(required=false)String destinationCountry,@RequestParam(required=false)String destination,@RequestParam(required=false)Long channelId,@RequestParam(required=false)Long logisticsChannelId,@RequestParam(required=false)String sfTrackingNo,@RequestParam(required=false)String trackingNo,@RequestParam(required=false)LocalDateTime createdFrom,@RequestParam(required=false)LocalDateTime createdTo,@RequestParam(defaultValue="1")int page,@RequestParam(defaultValue="20")int pageSize,@RequestParam(defaultValue="createdAt")String sortBy,@RequestParam(defaultValue="DESC")String sortDirection,@AuthenticationPrincipal Jwt jwt){if(workbenchFilter!=null&&!workbenchFilter.isBlank())return ApiResponse.success(query.listForCaller(tenant(jwt),user(jwt),status,workbenchFilter,storeId,page,pageSize));return ApiResponse.success(query.listFilteredForCaller(tenant(jwt),user(jwt),orderNo,storeId,status,one(destinationCountry,destination),one(channelId,logisticsChannelId),one(sfTrackingNo,trackingNo),createdFrom,createdTo,page,pageSize,sortBy,sortDirection));}
 @GetMapping("/export") public ResponseEntity<byte[]> export(@RequestParam(required=false)String orderNo,@RequestParam(required=false)Long storeId,@RequestParam(required=false)String status,@RequestParam(required=false)String destinationCountry,@RequestParam(required=false)String destination,@RequestParam(required=false)Long channelId,@RequestParam(required=false)Long logisticsChannelId,@RequestParam(required=false)String sfTrackingNo,@RequestParam(required=false)String trackingNo,@RequestParam(required=false)LocalDateTime createdFrom,@RequestParam(required=false)LocalDateTime createdTo,@RequestParam(defaultValue="createdAt")String sortBy,@RequestParam(defaultValue="DESC")String sortDirection,@RequestParam(required=false)List<Long> ids,@AuthenticationPrincipal Jwt jwt){if(exporter==null)throw new IllegalStateException("Order export is not configured");byte[] body=exporter.export(tenant(jwt),user(jwt),orderNo,storeId,status,one(destinationCountry,destination),one(channelId,logisticsChannelId),one(sfTrackingNo,trackingNo),createdFrom,createdTo,sortBy,sortDirection,ids);return ResponseEntity.ok().contentType(MediaType.parseMediaType("text/csv;charset=UTF-8")).header(HttpHeaders.CONTENT_DISPOSITION,"attachment; filename=orders.csv").body(body);}
 @GetMapping("/{orderId}") public ApiResponse<OrderDetailResponse> get(@PathVariable Long orderId,@AuthenticationPrincipal Jwt jwt){return ApiResponse.success(query.getDetailForCaller(tenant(jwt),user(jwt),orderId));}
 @PutMapping("/{orderId}") public ApiResponse<ShipmentOrderResponse> update(@PathVariable Long orderId,@Valid @RequestBody UpdateShipmentOrderRequest r,@RequestHeader(value="X-Request-Id",required=false)String requestId,@AuthenticationPrincipal Jwt jwt){return ApiResponse.success(lifecycle.updateDraft(tenant(jwt),user(jwt),orderId,r,requestId));}
 @PostMapping("/{orderId}/submit") public ApiResponse<ShipmentOrderResponse> submit(@PathVariable Long orderId,@Valid @RequestBody OrderActionRequest r,@RequestHeader(value="Idempotency-Key",required=false)String key,@RequestHeader(value="X-Request-Id",required=false)String requestId,@AuthenticationPrincipal Jwt jwt){return ApiResponse.success(lifecycle.submit(tenant(jwt),user(jwt),orderId,r,key,requestId));}
 @PostMapping("/{orderId}/cancel") public ApiResponse<ShipmentOrderResponse> cancel(@PathVariable Long orderId,@Valid @RequestBody OrderActionRequest r,@RequestHeader(value="Idempotency-Key",required=false)String key,@RequestHeader(value="X-Request-Id",required=false)String requestId,@AuthenticationPrincipal Jwt jwt){return ApiResponse.success(lifecycle.cancel(tenant(jwt),user(jwt),orderId,r,key,requestId));}
 private String one(String primary,String alias){if(primary!=null&&alias!=null&&!primary.equalsIgnoreCase(alias))throw new ShipmentOrderException("COMMON-1001",400);return primary!=null?primary:alias;} private Long one(Long primary,Long alias){if(primary!=null&&alias!=null&&!primary.equals(alias))throw new ShipmentOrderException("COMMON-1001",400);return primary!=null?primary:alias;} private Long tenant(Jwt j){try{return Long.valueOf(j.getClaimAsString("tenant_id"));}catch(Exception e){throw new ShipmentOrderException("COMMON-1004",403);}} private Long user(Jwt j){try{return Long.valueOf(j.getSubject());}catch(Exception e){throw new ShipmentOrderException("COMMON-1004",403);}}}