# openapi/shipflow-api.yaml
openapi: 3.0.3
info:
  title: ShipFlow API
  version: 1.0.0
  description: |
    ShipFlow 跨境物流履约与费用结算系统第一版 API 契约。
    所有租户业务接口从 JWT 身份上下文获取 tenant_id，不接受调用方任意传入。
servers:
  - url: /api/v1
    description: ShipFlow API v1
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: 分页查询当前租户物流订单
      operationId: listShipmentOrders
      description: |
        仅查询 JWT tenant_id 对应租户的订单，需要 `scope:TENANT`。
        status 按后端订单状态过滤；storeId 按当前租户店铺过滤。
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - name: workbenchFilter
          in: query
          description: 仅工作台使用；按当前调用方可见店铺过滤。
          schema: { type: string, enum: [PENDING_FEE_CONFIRMATION, MISSING_ADDRESS] }
        - name: storeId
          in: query
          schema: { $ref: '#/components/schemas/Id' }
        - { name: orderNo, in: query, schema: { type: string, maxLength: 64 } }
        - { name: destinationCountry, in: query, schema: { type: string, pattern: '^[A-Z]{2}$' } }
        - { name: destination, in: query, description: destinationCountry 的兼容别名, schema: { type: string, pattern: '^[A-Z]{2}$' } }
        - { name: channelId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - { name: logisticsChannelId, in: query, description: channelId 的兼容别名, schema: { $ref: '#/components/schemas/Id' } }
        - { name: trackingNo, in: query, schema: { type: string, maxLength: 128 } }
        - { name: sfTrackingNo, in: query, description: trackingNo 的兼容别名, schema: { type: string, maxLength: 128 } }
        - { name: createdFrom, in: query, schema: { $ref: '#/components/schemas/UtcDateTime' } }
        - { name: createdTo, in: query, schema: { $ref: '#/components/schemas/UtcDateTime' } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, orderNo, status, destinationCountry, channelId, estimatedFee], default: createdAt } }
        - { name: sortDirection, in: query, schema: { type: string, enum: [ASC, DESC], default: DESC } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 当前租户订单分页
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessShipmentOrderPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /orders/export:
    get:
      tags: [Orders]
      summary: 导出当前调用方可见订单
      operationId: exportShipmentOrders
      parameters:
        - { name: ids, in: query, schema: { type: array, items: { $ref: '#/components/schemas/Id' }, maxItems: 1000 } }
        - { name: orderNo, in: query, schema: { type: string, maxLength: 64 } }
        - { name: storeId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - { name: destinationCountry, in: query, schema: { type: string, pattern: '^[A-Z]{2}$' } }
        - { name: channelId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - { name: trackingNo, in: query, schema: { type: string, maxLength: 128 } }
        - { name: createdFrom, in: query, schema: { $ref: '#/components/schemas/UtcDateTime' } }
        - { name: createdTo, in: query, schema: { $ref: '#/components/schemas/UtcDateTime' } }
      responses:
        '200': { description: UTF-8 BOM CSV, content: { text/csv: { schema: { type: string, format: binary } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: 查询订单详情
      operationId: getOrder
      description: 仅查询 JWT tenant_id 对应租户的订单，需要 `scope:TENANT`。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: 订单详情
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessShipmentOrder'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Orders]
      summary: 更新草稿订单的非计价信息
      operationId: updateDraftShipmentOrder
      description: |
        仅允许当前租户的 DRAFT 订单更新收件地址、寄件地址、商品明细和备注，需要 `scope:TENANT`。
        渠道、目的国家、申报重量尺寸、包裹数、价格规则版本、计费重量、金额、币种和报价快照均禁止修改；传入这些字段返回 422。
        version 是必填字段并用于订单乐观锁；版本过期或订单已不再是 DRAFT 时返回 409。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShipmentOrderRequest'
      responses:
        '200': { description: 更新后的草稿订单, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessShipmentOrder' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/OrderRuleError' }
  /orders/{orderId}/submit:
    post:
      tags: [Orders]
      summary: 提交订单
      operationId: submitOrder
      description: |
        仅允许 DRAFT 进入 PENDING_INBOUND。
        提交请求必须携带 version，并使用订单乐观锁。
        DRAFT 不允许直接执行入库操作。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderActionRequest'
      responses:
        '200':
          description: 提交成功，状态为 PENDING_INBOUND
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessShipmentOrder'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/OrderRuleError'
  /orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: 取消草稿或待入库订单
      operationId: cancelOrder
      parameters: [{ $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderActionRequest' } } } }
      responses:
        '200': { description: 已取消订单, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessShipmentOrder' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /orders/{orderId}/tracking-events:
    get:
      tags: [Tracking]
      summary: 查询订单物流轨迹
      operationId: listTrackingEvents
      description: |
        仅返回 TrackingEventView，不返回 rawPayload、物流商签名或内部处理堆栈。
        按 eventTime 升序展示，迟到事件不得使订单状态回退。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: 轨迹分页
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTrackingPage'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /warehouse/orders/{orderId}:
    get:
      tags: [Warehouse]
      summary: 查询仓库作业详情
      operationId: getWarehouseWork
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      responses:
        '200': { description: 仓库作业详情, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessWarehouseWorkItem' } } } }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: REFRESH_TOKEN
      description: HttpOnly Refresh Token Cookie；客户端不可读取。
    providerHmac:
      type: apiKey
      in: header
      name: X-Provider-Signature
      description: Base64 编码的 HMAC-SHA256 签名，签名原文为 timestamp + "." + rawRequestBody。
  headers:
    TraceId:
      description: 与响应体 traceId 一致的链路追踪 ID。
      schema:
        type: string
    NoStore:
      description: 禁止缓存认证凭证相关响应。
      schema:
        type: string
        enum: [no-store]
        default: no-store
    WWWAuthenticate:
      description: RFC 6750 Bearer 认证挑战；具体错误详情不在 Header 中泄露敏感信息。
      schema:
        type: string
        example: Bearer
...
  parameters:
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 请求参数或格式错误。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AuthError:
      description: 未登录、Token 无效或已过期。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Forbidden:
      description: 权限不足或跨租户访问。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 资源不存在。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 幂等、乐观锁或订单状态冲突。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    OrderRuleError:
      description: 订单业务规则校验失败。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    Id:
      type: string
      pattern: '^[0-9]+$'
      example: '10001'
    Money:
      type: string
      pattern: '^[0-9]+\.[0-9]{2}$'
      example: '58.00'
      description: 非负金额，JSON 中使用字符串。
...
    CountryCode:
      type: string
      pattern: '^[A-Z]{2}$'
      example: JP
      description: 大写 ISO 3166-1 alpha-2 国家编码。
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      example: USD
      description: 大写 ISO 4217 三位币种编码。
    Weight:
      type: string
      pattern: '^[0-9]+\.[0-9]{3}$'
      example: '1.500'
      description: 重量，统一使用 kg，JSON 中使用字符串。
...
    UtcDateTime:
      type: string
      format: date-time
      example: '2026-08-02T10:30:00.000Z'
      description: UTC ISO-8601 时间。
...
    ApiSuccess:
      type: object
      required: [success, traceId, data]
      properties:
        success:
          type: boolean
          enum: [true]
        traceId:
          type: string
        message:
          type: string
          example: OK
        data: {}
...
    ApiError:
      type: object
      required: [success, traceId, error]
      properties:
        success:
          type: boolean
          enum: [false]
        traceId:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: ORDER-1001
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    PageMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    Address:
      type: object
      required: [contactName, phone, countryCode, city, addressLine1, postalCode]
      properties:
        contactName:
          type: string
        companyName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        countryCode:
          $ref: '#/components/schemas/CountryCode'
        stateProvince:
          type: string
        city:
          type: string
        district:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        postalCode:
          type: string
...
    ShipmentOrderItem:
      type: object
      additionalProperties: false
      required: [sku, productName, quantity, unitPrice, currency]
      properties:
        sku: { type: string, minLength: 1 }
        productName: { type: string, minLength: 1 }
        quantity: { type: number, format: decimal, exclusiveMinimum: true }
        unitPrice: { $ref: '#/components/schemas/Money' }
        currency: { type: string, minLength: 1 }
        hsCode: { type: string }
        countryOfOrigin: { type: string }
    UpdateShipmentOrderRequest:
      type: object
      additionalProperties: false
      required: [version, senderAddress, receiverAddress, items]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
        senderAddress: { $ref: '#/components/schemas/Address' }
        receiverAddress: { $ref: '#/components/schemas/Address' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ShipmentOrderItem' }
        remark: { type: string }
        channelId: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        destinationCountry: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredWeight: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredLength: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredWidth: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredHeight: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        packageCount: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        priceRuleVersion: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        chargeableWeight: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        amount: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        currency: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        quoteSnapshot: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
    ShipmentOrder:
      type: object
      required: [id, orderNo, quoteId, status, estimatedFee, currency, chargeableWeight, version, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        orderNo: { type: string }
        quoteId: { $ref: '#/components/schemas/Id' }
        status: { $ref: '#/components/schemas/OrderStatus' }
        estimatedFee: { $ref: '#/components/schemas/Money' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        chargeableWeight: { $ref: '#/components/schemas/Weight' }
        version:
          type: integer
          format: int64
          minimum: 0
          description: 当前订单乐观锁版本；草稿更新、提交或取消成功后返回递增后的最新值。
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
    ApiSuccessShipmentOrder:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ShipmentOrder' }
    ShipmentOrderPage:
      type: object
      required: [page, pageSize, totalPages, total, items]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        totalPages: { type: integer, format: int64, minimum: 0 }
        total: { type: integer, format: int64, minimum: 0 }
        items: { type: array, items: { $ref: '#/components/schemas/ShipmentOrder' } }
    ApiSuccessShipmentOrderPage:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ShipmentOrderPage' }
...
    OrderStatus:
      type: string
      enum: [DRAFT, PENDING_INBOUND, INBOUND, PENDING_PRICE_CONFIRMATION, READY_FOR_OUTBOUND, OUTBOUND, IN_TRANSIT, DELIVERED, CANCELLED, RETURNED, LOST]
...
    TrackingEventView:
      type: object
      required: [id, trackingNo, eventId, eventCode, eventTime, receivedTime, processStatus]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        trackingNo:
          type: string
        eventId:
          type: string
        eventCode:
          type: string
        eventDescription:
          type: string
        eventTime:
          $ref: '#/components/schemas/UtcDateTime'
        receivedTime:
          $ref: '#/components/schemas/UtcDateTime'
        processStatus:
          type: string
          enum: [PENDING, PROCESSED, RETRY, REJECTED]
    TrackingPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/TrackingEventView'
    ApiSuccessTrackingPage:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TrackingPage'
...
    OrderActionRequest:
      type: object
      required: [version]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
        reason: { type: string }
...
    WarehouseWorkItem:
      type: object
      required: [id, businessOrderNo, tenantId, tenantName, destinationCountry, chargeableWeight, estimatedFee, currentFee, currency, feeDifference, feeAlert, warehouseStatus, logisticsStatus, version, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        businessOrderNo: { type: string }
        sfTrackingNo: { type: string, nullable: true }
        tenantId: { $ref: '#/components/schemas/Id' }
        tenantName: { type: string }
        destinationCountry: { type: string }
        declaredWeight: { type: number, format: decimal }
        declaredLength: { type: number, format: decimal }
        declaredWidth: { type: number, format: decimal }
        declaredHeight: { type: number, format: decimal }
        declaredVolumeWeight: { type: number, format: decimal }
        actualWeight: { type: number, format: decimal, nullable: true }
        actualLength: { type: number, format: decimal, nullable: true }
        actualWidth: { type: number, format: decimal, nullable: true }
        actualHeight: { type: number, format: decimal, nullable: true }
        actualVolumeWeight: { type: number, format: decimal, nullable: true }
        chargeableWeight: { type: number, format: decimal }
        estimatedFee: { type: number, format: decimal }
        currentFee: { type: number, format: decimal }
        currency: { type: string }
        feeDifference: { type: number, format: decimal }
        feeAlert: { type: boolean }
        warehouseStatus: { type: string }
        logisticsStatus: { type: string }
        version: { type: integer, format: int64 }
        outboundAt: { $ref: '#/components/schemas/UtcDateTime', nullable: true }
        outboundBy: { $ref: '#/components/schemas/Id', nullable: true }
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
...
    ApiSuccessWarehouseWorkItem:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/WarehouseWorkItem' } } }]
...